Utilize array length index for RefExistsQuery
